# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WireGuard Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  username: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface SetupStatus {
  setupRequired: boolean;
}

export interface SetupInput {
  username: string;
  password: string;
}

export interface DashboardSummary {
  totalNodes: number;
  onlineNodes: number;
  totalClients: number;
  nodesWithWireguard: number;
}

export type NodeSshAuthMethod = typeof NodeSshAuthMethod[keyof typeof NodeSshAuthMethod];


export const NodeSshAuthMethod = {
  password: 'password',
  private_key: 'private_key',
} as const;

export type NodeWireguardStatus = typeof NodeWireguardStatus[keyof typeof NodeWireguardStatus];


export const NodeWireguardStatus = {
  not_installed: 'not_installed',
  stopped: 'stopped',
  running: 'running',
  unknown: 'unknown',
} as const;

export interface Node {
  id: number;
  name: string;
  ipAddress: string;
  sshUser: string;
  sshPort: number;
  sshAuthMethod: NodeSshAuthMethod;
  wireguardPort: number;
  wireguardStatus: NodeWireguardStatus;
  wireguardInterface: string;
  clientCount: number;
  createdAt: string;
}

export type NodeInputSshAuthMethod = typeof NodeInputSshAuthMethod[keyof typeof NodeInputSshAuthMethod];


export const NodeInputSshAuthMethod = {
  password: 'password',
  private_key: 'private_key',
} as const;

export interface NodeInput {
  name: string;
  ipAddress: string;
  sshUser: string;
  sshPort?: number;
  sshAuthMethod?: NodeInputSshAuthMethod;
  sshPassword?: string;
  sshPrivateKey?: string;
}

export type NodeSshUpdateInputSshAuthMethod = typeof NodeSshUpdateInputSshAuthMethod[keyof typeof NodeSshUpdateInputSshAuthMethod];


export const NodeSshUpdateInputSshAuthMethod = {
  password: 'password',
  private_key: 'private_key',
} as const;

export interface NodeSshUpdateInput {
  sshAuthMethod?: NodeSshUpdateInputSshAuthMethod;
  sshUser?: string;
  sshPort?: number;
  sshPassword?: string;
  sshPrivateKey?: string;
}

export interface WireguardSetupInput {
  port: number;
  interface?: string;
  subnet?: string;
}

export interface ConnectionTestResult {
  success: boolean;
  message: string;
  /** @nullable */
  latencyMs?: number | null;
}

export interface ImportResult {
  success: boolean;
  message: string;
  /** @nullable */
  wireguardInterface?: string | null;
  /** @nullable */
  wireguardPort?: number | null;
  /** @nullable */
  wireguardSubnet?: string | null;
  /** @nullable */
  wireguardStatus?: string | null;
  /** @nullable */
  existingPeerCount?: number | null;
}

export interface ActionResult {
  success: boolean;
  message: string;
  /** @nullable */
  output?: string | null;
}

export interface Client {
  id: number;
  nodeId: number;
  name: string;
  publicKey: string;
  allowedIps: string;
  createdAt: string;
}

export interface ClientInput {
  name: string;
}

export interface ClientConfig {
  config: string;
}

export interface ClientQr {
  qrBase64: string;
}