# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * WireGuard Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Client {
  id: number;
  nodeId: number;
  name: string;
  publicKey: string;
  allowedIps: string;
  createdAt: Date;
}